requirements.txt
amqp==2.5.2
asgiref==3.2.3
backports.csv==1.0.7
billiard==3.6.1.0
cairocffi==0.9.0
CairoSVG==2.1.3
celery==4.3.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.11.20
django-allauth==0.32.0
django-crispy-forms==1.8.0
django-filter==2.2.0
django-import-export==1.2.0
django-widget-tweaks==1.4.5
gunicorn==20.0.4
idna==2.8
importlib-metadata==0.23
kombu==4.6.6
more-itertools==7.2.0
oauthlib==3.1.0
Pillow==6.1.0
psycopg2-binary==2.8.4
pycparser==2.18
python-http-client==3.2.1
python3-openid==3.1.0
pytz==2018.5
requests==2.22.0
requests-oauthlib==1.3.0
sendgrid-django==4.2.0
sqlparse==0.3.0
tablib==1.0.0
tinycss2==0.6.1
urllib3==1.25.8
vine==1.3.0
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
